Solidity verifier development continued (#103)
* Added public input evals

* Fixes

* Absorb public evals

* Added combined_evals and others

* Added get alphas

* Minor

* Progress on ft eval

* Added combined inner prod

* Fixing comb inn prod

* Fixed 3d array prob

* Fix compile errors:

* Add error

* README

* Deserialize VerifierIndex from KZG prover and fixes (#89)

* Add foreign group usage

* KZG Prover generate verifier index

* [WIP] Debug ForeignGroup assertEquals

* Replaced job's OS. (#88)

* Use foreign EC addition gate

* Deserialize verifier index

* Added remaining msgpack primitives

* Go back to min Kimchi verifier

* Use foreign EC op gates

* Fix uint deser

* Remove unused imports

* Update o1js

* Add public_len field

* Add ForeignFieldMul gate case

* Initialize alphas

* Kimchi o1js prover: combine_point_evaluations function deleted (#91)


* Final fixes

---------


* Remove repeated lines


* Remove unused import


* Change o1js submodule commit

* Changed commit again

* Ethereum verifier continued (#92)

* KZG Prover generate verifier index

* Deserialize verifier index

* Added remaining msgpack primitives

* Fix uint deser

* Add public_len field

* Initialize alphas

* Final fixes

* Add constraint count for pows of alpha

* Added combined evaluations step

* Added fake witness

* Fixed issues regarding fake witness

* Added comments, removed printlns

* Satisfy copy constraints

* Revert fake witness development

* Added test circuit kzg proof

* Disable test circuit proof gen

* Added vanishing poly

* Continuing final comm step

* Added perm scalars

* Added PolishToken

* Added new test circuit for the KZG prover

* Added better comment to polishtoken

* RPN progress

* Finished preliminary rpn evaluation

* Fixes

* Test verifier index from new circuit

* WIP fixes for deserialization

* Final fixes

* Deserialized part of the prover proof

* Deserialize public evals

* Deserialize rest of evals

* Finish deserializing index and proof

* Remove duplicated function

---------

+++ eth_verifier/lib/VerifierIndex.sol
@@ -16,15 +16,14 @@ struct VerifierIndex {
     uint64 zk_rows;
     URS urs;
     // domain
-    uint256 domain_size;
+    uint64 domain_size;
     Scalar.FE domain_gen;
     /// The mapping between powers of alpha and constraints
     Alphas powers_of_alpha;
     // wire shift coordinates
-    Scalar.FE[7] shift;  // TODO: use Consants.PERMUTS
+    Scalar.FE[7] shift; // TODO: use Consants.PERMUTS
     /// domain offset for zero-knowledge
     Scalar.FE w;
-
 }
 
 function verifier_digest(VerifierIndex storage index) returns (Base.FE) {
